requirements.txt
Flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.1
flask-mail==0.9.1
flask-cors==4.0.0
flask-migrate==4.0.5
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
gunicorn==21.2.0
email-validator==2.1.0.post1
reportlab==4.0.7
Pillow==10.1.0
setuptools>=68.0.0
wheel>=0.40.0
